Backend/dealsduka/requirements.txt
Django==4.2.25
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.0.0
Pillow==10.4.0
gunicorn==21.2.0
drf-yasg==1.21.7

dj-database-url==1.0.0
whitenoise==6.5.0
drf-yasg
setuptools


